Create a seperate file prepare regard application
