+++ server/service/IndividualServicesService.js
@@ -6,7 +6,6 @@ const LogicalTerminationPointServiceOfUtility = require("onf-core-model-ap-bs/ba
 const ForwardingConfigurationService = require('onf-core-model-ap/applicationPattern/onfModel/services/ForwardingConstructConfigurationServices');
 const ForwardingAutomationService = require('onf-core-model-ap/applicationPattern/onfModel/services/ForwardingConstructAutomationServices');
 const prepareForwardingConfiguration = require('./individualServices/PrepareForwardingConfiguration');
-const prepareForwardingAutomation = require('./individualServices/PrepareForwardingAutomation');
 const prepareALTForwardingAutomation = require('onf-core-model-ap-bs/basicServices/services/PrepareALTForwardingAutomation');
 
 const httpClientInterface = require('onf-core-model-ap/applicationPattern/onfModel/models/layerProtocols/HttpClientInterface');
@@ -23,7 +22,7 @@ const softwareUpgrade = require('./individualServices/SoftwareUpgrade');
 const { getIndexAliasAsync, createResultArray, elasticsearchService } = require('onf-core-model-ap/applicationPattern/services/ElasticsearchService');
 const individualServicesOperationsMapping = require('./individualServices/IndividualServicesOperationsMapping');
 const TcpObject = require('onf-core-model-ap/applicationPattern/onfModel/services/models/TcpObject');
-
+const PrepareRegardApplication = require('./individualServices/PrepareRegardApplication');
 const REDIRECT_SERVICE_REQUEST_OPERATION = '/v1/redirect-service-request-information';
 const NEW_RELEASE_FORWARDING_NAME = 'PromptForBequeathingDataCausesTransferOfListOfApplications';
 
@@ -401,7 +400,7 @@ exports.regardApplication = async function (body, user, originator, xCorrelator,
         customerJourney
       );
       
-      let result = await prepareForwardingAutomation.regardApplication(
+      let result = await PrepareRegardApplication.regardApplication(
         applicationName,
         releaseNumber,
         user,
